derep skbio (#58)
This worked really well. The memory improvements showed. When I ran this on V4 ref data using the original implementation, the RAM usage spiked ~ 1 GB, and hovered ~ 880 MB. With the new implementation in this PR, the RAM usage spiked to half, ~ 500 MB, and hovered at ~ 330 MB. I suspect the RAM savings will grow with larger files.